bionic: Allow devices to add device specfic static libs
This provides a convenient (albeit powerful enough to be dangerous)
hook to add symbols needed to support vendor blobs that do not
necessarily match our source code.

Doing so via this hook has several advantages over patching the
code in question:

* The hacks do no pollute other repositories
* The hacks do not have any risk of breaking any other devices
* The hacks don't just live forever when we forget they exist
* The hacks are all easy to find by locating them together

When using this, please take extra care to include only the
minimal code to support the change.  Keep in mind that all
code (and all libraries your code links against) will be
part of the address space of every single process in
the system!

Change-Id: I6dcd9ad7cee330febe1a974619144d378b67b364